requirements.txt
django==1.11.8
djangorestframework==3.7.7
django-mptt==0.9.0
djangorestframework-recursive==0.1.2